% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SankGplot.R
\name{SankeyGplot}
\alias{SankeyGplot}
\title{Load a Matrix}
\usage{
SankeyGplot(fileDir)
}
\arguments{
\item{filDir}{Path to the input csv file}
}
\value{
A Sankey Diagram of input file
}
\description{
This function loads a file directory to plot the sankey diagram using ggplot library.
It assumes the input file is a csv file with the columns being the risk factor names and
the corresponding years in which the measurements are recorded.
In order to have a standard form, the column names are changed into Risk Factor, Year and Measurement
kepted.
}
\keyword{SankeyGPlot}
